requirements.txt
openai>=0.27.0
python-dotenv>=0.21.0
requests>=2.28.0
pytest>=7.0.0
reflex==0.8.24.post1
